requirements.txt
python-telegram-bot
flask
psutil
aiosqlite
httpx
